+++ src/runner.ts
@@ -2,41 +2,33 @@ import { spawnSync } from "child_process";
 import * as vscode from 'vscode';
 
 export function runOnFile(filename:string, workspace:string, config: {[key:string]: any}){
-    let result = runCppLint(filename, workspace, config);
+    let result = runCppLint(filename, workspace, config, false);
     return result;
 }
 
 export function runOnWorkspace(filename:string, workspace:string, config: {[key:string]: any}){
-    let result = runWholeCppLint(filename, workspace, config);
+    let result = runCppLint(filename, workspace, config, true);
     return result;
 }
 
-function runCppLint(filename:string, workspace:string, config: {[key:string]: any}) {
+function runCppLint(filename:string, workspace:string, config: {[key:string]: any}, enableworkspace:boolean) {
     let start = 'CppLint started: ' + new Date().toString();
     let cpplint = config["cpplintPath"];
-    let linelength = "--linelength=" + config['lineLength']
-    let param:string[] = ['--output=vs7']
+    let linelength = "--linelength=" + config['lineLength'];
+    let param:string[] = ['--output=vs7', linelength];
     config['excludes'].forEach(element => {
         param.push("--exclude=" + element)
     });
-    param = param.concat([linelength, filename]);
-    let result = spawnSync(cpplint, param, {'cwd': workspace})
-    let stdout = '' + result.stdout;
-    let stderr = '' + result.stderr;
-    let end = 'CppLint ended: ' + new Date().toString();
-    let out = [start, stdout, stderr, end].join('\n');
-    return out;
-}
+    config['filters'].forEach(element => {
+        param.push("--filter=" + element)
+    });
+    param.push("--verbose=" + config['verbose']);
 
-function runWholeCppLint(filename:string, workspace:string, config: {[key:string]: any}) {
-    let start = 'CppLint started: ' + new Date().toString();
-    let cpplint = config["cpplintPath"];
-    let linelength = "--linelength=" + config['lineLength']
-    let param:string[] = ['--output=vs7']
-    config['excludes'].forEach(element => {
-        param.push("--exclude=" + element)
-    });
-    param = param.concat([linelength, "--recursive", "."]);
+    if (enableworkspace) {
+        param = param.concat([ "--recursive", "."]);
+    } else {
+        param.push(filename);
+    }
     let result = spawnSync(cpplint, param, {'cwd': workspace})
     let stdout = '' + result.stdout;
     let stderr = '' + result.stderr;